Move artefact builds controllers out of artefacts module
